test-requirements.txt
django==1.11.15
gunicorn
selenium
fabric3
